moved get odometer to test result repository
