+++ src/services/CertificateGenerationService.ts
@@ -16,7 +16,6 @@ import {
 	IRequiredStandard,
 	IRoadworthinessCertificateData,
 	ITestResult,
-	ITestType,
 	IWeightDetails,
 } from '../models';
 import {
@@ -41,6 +40,7 @@ import { IFlatDefect } from '../models/IFlatDefect';
 import { IItem } from '../models/IItem';
 import { ISearchResult, TechRecordGet, TechRecordType } from '../models/Types';
 import { TechRecordRepository } from '../tech-record/TechRecordRepository';
+import { TestResultRepository } from '../test-result/TestResultRepository';
 import { TestStationRepository } from '../test-station/TestStationRepository';
 import { TrailerRepository } from '../trailer/TrailerRepository';
 import { Configuration } from '../utils/Configuration';
@@ -59,7 +59,8 @@ class CertificateGenerationService {
 		private lambdaClient: LambdaService,
 		private trailerRepository: TrailerRepository,
 		private techRecordRepository: TechRecordRepository,
-		private testStationRepository: TestStationRepository
+		private testStationRepository: TestStationRepository,
+		private testResultRepository: TestResultRepository
 	) {}
 
 	/**
@@ -349,7 +350,9 @@ class CertificateGenerationService {
 			payload.MSVA_DATA = { ...msvaData };
 		} else {
 			const odometerHistory =
-				vehicleType === VEHICLE_TYPES.TRL ? undefined : await this.getOdometerHistory(systemNumber);
+				vehicleType === VEHICLE_TYPES.TRL
+					? undefined
+					: await this.testResultRepository.getOdometerHistory(systemNumber);
 			const TrnObj = this.isValidForTrn(vehicleType, makeAndModel)
 				? await this.trailerRepository.getTrailerRegistrationObject(testResult.vin, makeAndModel.Make)
 				: undefined;
@@ -650,84 +653,6 @@ class CertificateGenerationService {
 			console.log('No techRecord found for weight details');
 			throw new HTTPError(500, 'No vehicle found for Roadworthiness test certificate!');
 		}
-	}
-
-	/**
-	 * Retrieves the odometer history for a given VIN from the Test Results microservice
-	 * @param systemNumber - systemNumber for which to retrieve odometer history
-	 */
-	public async getOdometerHistory(systemNumber: string) {
-		const config: IInvokeConfig = this.config.getInvokeConfig();
-		const invokeParams: InvocationRequest = {
-			FunctionName: config.functions.testResults.name,
-			InvocationType: 'RequestResponse',
-			LogType: 'Tail',
-			Payload: toUint8Array(
-				JSON.stringify({
-					httpMethod: 'GET',
-					path: `/test-results/${systemNumber}`,
-					pathParameters: {
-						systemNumber,
-					},
-				})
-			),
-		};
-
-		return this.lambdaClient
-			.invoke(invokeParams)
-			.then((response: InvocationResponse) => {
-				const payload: any = this.lambdaClient.validateInvocationResponse(response);
-				// TODO: convert to correct type
-				const testResults: any[] = JSON.parse(payload.body);
-
-				if (!testResults || testResults.length === 0) {
-					throw new HTTPError(400, `${ERRORS.LAMBDA_INVOCATION_BAD_DATA} ${JSON.stringify(payload)}.`);
-				}
-				// Sort results by testEndTimestamp
-				testResults.sort((first: any, second: any): number => {
-					if (moment(first.testEndTimestamp).isBefore(second.testEndTimestamp)) {
-						return 1;
-					}
-
-					if (moment(first.testEndTimestamp).isAfter(second.testEndTimestamp)) {
-						return -1;
-					}
-
-					return 0;
-				});
-
-				// Remove the first result as it should be the current one.
-				testResults.shift();
-
-				// Set the array to only submitted tests (exclude cancelled)
-				const submittedTests = testResults.filter((testResult) => {
-					return testResult.testStatus === 'submitted';
-				});
-
-				const filteredTestResults = submittedTests
-					.filter(({ testTypes }) =>
-						testTypes?.some(
-							(testType: ITestType) =>
-								testType.testTypeClassification === 'Annual With Certificate' &&
-								(testType.testResult === 'pass' || testType.testResult === 'prs')
-						)
-					)
-					.slice(0, 3); // Only last three entries are used for the history.
-
-				return {
-					OdometerHistoryList: filteredTestResults.map((testResult) => {
-						return {
-							value: testResult.odometerReading,
-							unit: testResult.odometerReadingUnits,
-							date: moment(testResult.testEndTimestamp).format('DD.MM.YYYY'),
-						};
-					}),
-				};
-			})
-			.catch((error: ServiceException | Error) => {
-				console.log(error);
-				throw error;
-			});
 	}
 
 	/**
